biginteger: fix overflow error from 64-bit internal data structure.
Initially `unsigned long` type is used as an internal data structure and
the type was considered as a 32-bit unsigned integer.

However, The size of signed/unsigned `long` type depends on the architecture
on which the program is running. If the program is on a 64-bit
architecture system, it will be a 64-bit integer which is the same with
`unsigned long long`.

Therefore, the internal `unsigned long` structure on 64-bit systems
cannot detect overflow errors.

The internal `unsigned long` type is changed to `uint32_t` which is the
same with `unsigned int` to prevent overflows and to express its size
explicitly.