biginteger: eliminate wasted space from `_base`
biginteger stores the number onto a uint32_t vector.

In order to fit the number to each uint32_t element, it modulo the
number with `_base` which was the maximum number of 32-bit unsigned
integers.
After modulo operation, the number is limited by the maximum number - 1
even though the space can store the maximum number.

Therefore it is improved to store (maximum + 1) for each element.